fix fmt and add comments for upgrade
