+++ storage/src/table_info/legacy.rs
