+++ storage/src/upgrade.rs
@@ -28,7 +28,7 @@ pub static BARNARD_HARD_FORK_HASH: Lazy<HashValue> = Lazy::new(|| {
     HashValue::from_hex_literal(
         "0xf6fc5d0d737e0b9c5768a62a9b4b3bb79b9b1bc47c62fa9fb3b215157dbed9ac",
     )
-        .expect("")
+    .expect("")
 });
 
 pub static DRAGON_HARD_FORK_HEIGHT: BlockNumber = 16801958;
@@ -36,7 +36,7 @@ pub static DRAGON_HARD_FORK_HASH: Lazy<HashValue> = Lazy::new(|| {
     HashValue::from_hex_literal(
         "0xbef8d0af3b358af9fe25f7383fd2580679c54fe2ce7ff7a7434785ba6d11b943",
     )
-        .expect("")
+    .expect("")
 });
 
 impl DBUpgrade {
